+++ distributions/startupway/src/ui/index.ts
@@ -1,6 +1,6 @@
 import "@startupway/main/lib/ui";
+import "@startupway/workspace/lib/ui";
 import "@startupway/menu/lib/ui";
 import "@startupway/users/lib/ui";
-import "@startupway/workspace/lib/ui";
 // import { getUi } from "@startupway/main/lib/ui";
 // getUi().start();
