+++ packages/users/package.json
@@ -24,7 +24,6 @@
     "@startupway/main": "*",
     "@startupway/database": "*",
     "@startupway/menu": "*",
-    "@startupway/workspace": "*",
     "express": "^4.17.1"
   },
   "devDependencies": {
