+++ packages/users/src/ui/index.ts
@@ -2,36 +2,11 @@
 export * from "./ui";
 export * from "../common";
 import { UI } from '@startupway/main/lib/ui';
-import { WorkspaceUI, ToolbarButtonPosition } from '@startupway/workspace/lib/ui';
 import store from "./store";
 import { RouteConfig } from "vue-router";
 
 import Login from "./views/Login.vue";
-import UserMenu from "./views/UserMenu.vue";
-import EditAccount from "./views/EditAccount.vue";
-import EditSecuritySettings from "./views/EditSecuritySettings.vue";
 
-const workspaceUi = WorkspaceUI.getInstance()	;
-
-const workspaceRoutes: RouteConfig[] = [
-	{
-		path: "/user/security",
-		name:"Edit Security Settings",
-		component: EditSecuritySettings
-	},
-	{
-		path: "/user/account",
-		name:"Edit Account",
-		component: EditAccount
-	}
-]
-
-workspaceUi.registerToolbarButton (UserMenu, {
-	position: ToolbarButtonPosition.RIGHT,
-	priority: 3
-});
-
-workspaceUi.registerWorkspaceRoutes(workspaceRoutes);
 
 const ui = UI.getInstance();
