+++ packages/workspace/src/ui/index.ts
@@ -3,19 +3,39 @@ export * from "./ui";
 export * from "../common/common";
 import { UI } from '@startupway/main/lib/ui';
 import { WorkspaceUI } from './ui';
+import { ToolbarButtonPosition } from "../common/common";
 import store from "./store";
 import { RouteConfig } from "vue-router";
+
 import Workspace from "./views/Workspace.vue";
 import Recovery from "./views/Recovery.vue";
+import UserMenu from "./views/UserMenu.vue";
+import EditAccount from "./views/EditAccount.vue";
+import EditSecuritySettings from "./views/EditSecuritySettings.vue";
 
 const ui = UI.getInstance();
-const workspaceUI: WorkspaceUI = WorkspaceUI.getInstance();
+const workspaceUi = WorkspaceUI.getInstance();
+
+const workspaceRoutes: RouteConfig[] = [
+	{
+		path: "/user/security",
+		name:"Edit Security Settings",
+		component: EditSecuritySettings
+	},
+	{
+		path: "/user/account",
+		name:"Edit Account",
+		component: EditAccount
+	}
+]
+
+workspaceUi.addWorkspaceRoutes(workspaceRoutes);
 
 const routes: RouteConfig[] = [
 	{
 		name:"workspace",
 		path: '/workspace',
-		children: workspaceUI.routes,
+		children: workspaceUi.routes,
 		component: Workspace
 	},
 	{
@@ -28,3 +48,8 @@ const routes: RouteConfig[] = [
 
 ui.registerRoutes(routes);
 ui.registerStore("workspace",store());
+
+workspaceUi.registerToolbarButton (UserMenu, {
+	position: ToolbarButtonPosition.RIGHT,
+	priority: 3
+});
